fix typo in vignette link; update impact page (#16)
Fixed typo in link to definitions vignette

Regenerated the "impact" page, so the numbers are up to date & edited how the parameters "digits" is used, because it was failing to rebuild on those lines of code.

Note: The line of code in assets/jsconfig.json was changed automatically. I assume this is correct.

Note: When I regenerated the webpage on my machine, all the twitter images opened as separate windows, instead of being embedded in the page. I'm assuming this was something that will happen locally only.